Extract Unix fn from self_update.rs
This introduces a strategy of defining a module that describes
OS-specific functions and re-exports those functions from there.
